refactor: split AmplificationHelper to several little classes with clearer purpose
