+++ dspot/src/main/java/eu/stamp_project/dspot/amplifier/AbstractAmplifier.java
@@ -1,6 +1,6 @@
 package eu.stamp_project.dspot.amplifier;
 
-import eu.stamp_project.utils.AmplificationHelper;
+import eu.stamp_project.utils.CloneHelper;
 import eu.stamp_project.utils.Counter;
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.declaration.CtMethod;
@@ -71,7 +71,7 @@ public abstract class AbstractAmplifier<T extends CtElement> implements Amplifie
     protected CtMethod<?> replace(T originalElement, T amplifiedElement, CtMethod<?> testMethod) {
         originalElement.replace(amplifiedElement);
         amplifiedElement.putMetadata(this.METADATA_KEY, true);
-        CtMethod<?> clone = AmplificationHelper.cloneTestMethodForAmp(testMethod, getSuffix());
+        CtMethod<?> clone = CloneHelper.cloneTestMethodForAmp(testMethod, getSuffix());
         amplifiedElement.replace(originalElement);
         Counter.updateInputOf(clone, 1);
         return clone;
