+++ dspot/src/test/java/eu/stamp_project/AbstractTest.java
@@ -1,7 +1,7 @@
 package eu.stamp_project;
 
 import eu.stamp_project.dspot.amplifier.value.ValueCreator;
-import eu.stamp_project.utils.AmplificationHelper;
+import eu.stamp_project.utils.RandomHelper;
 import org.junit.Before;
 
 /**
@@ -18,7 +18,7 @@ public abstract class AbstractTest {
     @Before
     public void setUp() throws Exception {
         Utils.init(getPathToPropertiesFile());
-        AmplificationHelper.setSeedRandom(72L);
+        RandomHelper.setSeedRandom(72L);
         ValueCreator.count = 0;
         Utils.getInputConfiguration().setVerbose(true);
         Utils.getInputConfiguration().setMinimize(false);
