+++ dspot/src/test/java/eu/stamp_project/dspot/amplifier/BooleanLiteralAmplifierTest.java
@@ -2,7 +2,7 @@ package eu.stamp_project.dspot.amplifier;
 
 import eu.stamp_project.Utils;
 import eu.stamp_project.AbstractTest;
-import eu.stamp_project.utils.AmplificationHelper;
+import eu.stamp_project.utils.RandomHelper;
 import org.junit.Test;
 import spoon.reflect.declaration.CtClass;
 import spoon.reflect.declaration.CtMethod;
@@ -18,7 +18,7 @@ public class BooleanLiteralAmplifierTest extends AbstractTest {
     public void testAmplify() throws Exception {
         final String nameMethod = "methodBoolean";
         CtClass<Object> literalMutationClass = Utils.getFactory().Class().get("fr.inria.amp.LiteralMutation");
-        AmplificationHelper.setSeedRandom(42L);
+        RandomHelper.setSeedRandom(42L);
         Amplifier mutator = new BooleanLiteralAmplifier();
         CtMethod method = literalMutationClass.getMethod(nameMethod);
         List<CtMethod> mutantMethods = mutator.amplify(method, 0).collect(Collectors.toList());
