+++ dspot/src/test/java/eu/stamp_project/dspot/amplifier/ReplacementAmplifierTest.java
@@ -3,6 +3,7 @@ package eu.stamp_project.dspot.amplifier;
 import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
 import eu.stamp_project.utils.AmplificationHelper;
+import eu.stamp_project.utils.RandomHelper;
 import org.junit.Test;
 import spoon.reflect.declaration.CtMethod;
 import spoon.reflect.factory.Factory;
@@ -24,7 +25,7 @@ public class ReplacementAmplifierTest extends AbstractTest {
 
         final String packageName = "fr.inria.statementadd";
         final Factory factory = Utils.getFactory();
-        AmplificationHelper.setSeedRandom(32L);
+        RandomHelper.setSeedRandom(32L);
         ReplacementAmplifier amplifier = new ReplacementAmplifier();
         amplifier.reset(factory.Class().get(packageName + ".ClassTarget"));
 
@@ -50,7 +51,7 @@ public class ReplacementAmplifierTest extends AbstractTest {
 
         final String packageName = "fr.inria.statementadd";
         final Factory factory = Utils.getFactory();
-        AmplificationHelper.setSeedRandom(32L);
+        RandomHelper.setSeedRandom(32L);
         ReplacementAmplifier amplifier = new ReplacementAmplifier();
         amplifier.reset(factory.Class().get(packageName + ".ClassTarget"));
